Move armor function defs to character_armor.cpp (#51369)
Including:
- Character::can_interface_armor
- Character::mutation_armor (x3)
- Character::get_armor_bash
- Character::get_armor_cut
- Character::get_armor_bullet
- Character::get_armor_type
- Character::get_all_armor_type
- Character::get_armor_bash_base
- Character::get_armor_cut_base
- Character::get_armor_bullet_base
- Character::get_armor_acid
- Character::get_env_resist
- Character::get_armor_fire
- static destroyed_armor_msg
- static item_armor_enchantment_adjust
- static armor_enchantment_adjust
- Character::absorb_hit
- Character::armor_absorb
